SoOrganizedApp/requirements.txt
asgiref==3.6.0
Django==4.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
